Corrections

+++ session-2016-04-idf/coefs/mcx_framework.py
@@ -34,7 +34,7 @@ erreur = HEADER + """
 void erreur( int bits, double exact, double approx, int resolution )
  {
   if (exact==0) { echec(1,"division par 0") ; }
-  int erreur = arrondi(resolution*double(exact-approx)/exact) ;
+  int erreur = arrondi(resolution*(exact-approx)/exact) ;
   if (erreur<0) { erreur = -erreur ; }
   std::cout
     <<std::right<<std::setw(2)<<bits<<" bits : "
@@ -58,7 +58,7 @@ class Testeur
     void erreur( int bits, double exact, double approx )
      {
       if (exact==0) { echec(1,"division par 0") ; }
-      int erreur = arrondi(resolution_*double(exact-approx)/exact) ;
+      int erreur = arrondi(resolution_*(exact-approx)/exact) ;
       if (erreur<0) { erreur = -erreur ; }
       std::cout
         <<std::right<<std::setw(2)<<bits<<" bits : "
@@ -90,7 +90,7 @@ class Testeur
     void erreur( int bits, double exact, double approx )
      {
       if (exact==0) { echec(1,"division par 0") ; }
-      int erreur = arrondi(resolution_*double(exact-approx)/exact) ;
+      int erreur = arrondi(resolution_*(exact-approx)/exact) ;
       if (erreur<0) { erreur = -erreur ; }
       std::cout
         <<std::right<<std::setw(2)<<bits<<" bits : "
@@ -162,7 +162,7 @@ class Testeur
     void erreur( int bits, double exact, double approx )
      {
       if (exact==0) { echec(1,"division par 0") ; }
-      int erreur = arrondi(resolution_*double(exact-approx)/exact) ;
+      int erreur = arrondi(resolution_*(exact-approx)/exact) ;
       if (erreur<0) { erreur = -erreur ; }
       std::cout
         <<std::right<<std::setw(2)<<bits<<" bits : "
@@ -278,7 +278,7 @@ class Testeur
     void erreur( int bits, double exact, double approx )
      {
       if (exact==0) { echec(1,"division par 0") ; }
-      int erreur = arrondi(resolution_*double(exact-approx)/exact) ;
+      int erreur = arrondi(resolution_*(exact-approx)/exact) ;
       if (erreur<0) { erreur = -erreur ; }
       std::cout
         <<std::right<<std::setw(2)<<bits<<" bits : "
@@ -388,7 +388,7 @@ class Testeur
     void erreur( int bits, double exact, double approx )
      {
       if (exact==0) { echec(1,"division par 0") ; }
-      int erreur = arrondi(resolution_*double(exact-approx)/exact) ;
+      int erreur = arrondi(resolution_*(exact-approx)/exact) ;
       if (erreur<0) { erreur = -erreur ; }
       std::cout
         <<std::right<<std::setw(2)<<bits<<" bits : "
@@ -454,7 +454,7 @@ class Testeur
     void erreur( int bits, double exact, double approx )
      {
       if (exact==0) { echec(1,"division par 0") ; }
-      int erreur = arrondi(resolution_*double(exact-approx)/exact) ;
+      int erreur = arrondi(resolution_*(exact-approx)/exact) ;
       if (erreur<0) { erreur = -erreur ; }
       std::cout
         <<std::right<<std::setw(2)<<bits<<" bits : "
@@ -625,7 +625,7 @@ class Testeur
     void erreur( int bits, double exact, double approx )
      {
       if (exact==0) { throw Echec(1,"division par 0") ; }
-      int erreur = arrondi(resolution_*double(exact-approx)/exact) ;
+      int erreur = arrondi(resolution_*(exact-approx)/exact) ;
       if (erreur<0) { erreur = -erreur ; }
       std::cout
         <<std::right<<std::setw(2)<<bits<<" bits : "
@@ -777,7 +777,7 @@ class Testeur
     void erreur( int bits, double exact, double approx )
      {
       if (exact==0) { throw Echec(1,"division par 0") ; }
-      int erreur = arrondi(resolution_*double(exact-approx)/exact) ;
+      int erreur = arrondi(resolution_*(exact-approx)/exact) ;
       if (erreur<0) { erreur = -erreur ; }
       std::cout
         <<std::right<<std::setw(2)<<bits<<" bits : "
@@ -830,7 +830,7 @@ class Testeur
     void erreur( int bits, double exact, double approx )
      {
       if (exact==0) { throw EchecDivisionParZero() ; }
-      int erreur = arrondi(resolution_*double(exact-approx)/exact) ;
+      int erreur = arrondi(resolution_*(exact-approx)/exact) ;
       if (erreur<0) { erreur = -erreur ; }
       std::cout
         <<std::right<<std::setw(2)<<bits<<" bits : "
